requirements.txt
pandas
numpy
tqdm
torch
datasets
transformers
pinecone-client
nltk
rouge-score
bert-score